requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
